+++ src/main/java/me/shedaniel/rei/api/DisplayHelper.java
@@ -1,9 +1,12 @@
 package me.shedaniel.rei.api;
 
 import me.shedaniel.rei.RoughlyEnoughItemsCore;
+import net.minecraft.util.ActionResult;
 
 import java.awt.*;
 import java.util.List;
+
+import static net.minecraft.util.ActionResult.PASS;
 
 public interface DisplayHelper {
     
@@ -19,6 +22,10 @@ public interface DisplayHelper {
         Rectangle getLeftBounds(T screen);
         
         Rectangle getRightBounds(T screen);
+        
+        default ActionResult canItemSlotWidgetFit(boolean isOnRightSide, int left, int top, T screen, Rectangle fullBounds) {
+            return PASS;
+        }
         
         default Rectangle getItemListArea(Rectangle rectangle) {
             return new Rectangle(rectangle.x + 2, rectangle.y + 24, rectangle.width - 4, rectangle.height - (RoughlyEnoughItemsCore.getConfigManager().getConfig().sideSearchField ? 27 + 22 : 27));